[fix🚑] : ci 에러 수정 및 doc
